Add phaseManager.js module for improved modularity
- Extract phase management functions to separate module
- Implement ES6 module imports/exports
- Update main JS file to import from phaseManager.js
- Ensure all phase transitions and UI updates work correctly
- Maintain integration with gameState module
